[refactor] 폴리곤 좌표 리스트 common 모듈로 이동
:feature:booth 모듈과 :feature:map 모듈이 동일한 리스트를 공유하기 위함, BoothLocation 화면 내에도 여러 학교에 대한 폴리콘 좌표 적용
