+++ core/common/src/main/kotlin/com/unifest/android/core/common/SchoolPolylines.kt
@@ -1,0 +1,292 @@
+package com.unifest.android.core.common
+
+import com.naver.maps.geometry.LatLng
+import kotlinx.collections.immutable.persistentListOf
+
+val KONKUK_UNIVERSITY_POLYLINE = persistentListOf(
+    LatLng(37.0125281, 127.2598307),
+    LatLng(37.0101251, 127.2631513),
+    LatLng(37.0095553, 127.2639023),
+    LatLng(37.0094097, 127.2642242),
+    LatLng(37.0096453, 127.264546),
+    LatLng(37.0098295, 127.2649537),
+    LatLng(37.0100737, 127.2654687),
+    LatLng(37.0102578, 127.2656458),
+    LatLng(37.010502, 127.2659032),
+    LatLng(37.0112431, 127.2661661),
+    LatLng(37.0113802, 127.2661862),
+    LatLng(37.0114947, 127.2661782),
+    LatLng(37.0122818, 127.2659301),
+    LatLng(37.0125174, 127.2658979),
+    LatLng(37.0127305, 127.2659019),
+    LatLng(37.0130015, 127.2659703),
+    LatLng(37.0130464, 127.2657544),
+    LatLng(37.0131846, 127.2654915),
+    LatLng(37.0132178, 127.2654245),
+    LatLng(37.0133816, 127.2649189),
+    LatLng(37.0135626, 127.2649269),
+    LatLng(37.0136301, 127.2645246),
+    LatLng(37.013629, 127.2643958),
+    LatLng(37.0134652, 127.2634343),
+    LatLng(37.0141269, 127.2632143),
+    LatLng(37.0140359, 127.262694),
+    LatLng(37.0139781, 127.2626028),
+    LatLng(37.013266, 127.2617921),
+    LatLng(37.0133602, 127.2616184),
+    LatLng(37.0133709, 127.2615789),
+    LatLng(37.0128826, 127.2611242),
+    LatLng(37.0128044, 127.2610478),
+    LatLng(37.0131423, 127.2604087),
+    LatLng(37.0125265, 127.2598301),
+)
+
+val KOREA_NATIONAL_UNIVERSITY_OF_TRANSPORTATION_POLYLINE = persistentListOf(
+    LatLng(36.9665013, 127.8746701),
+    LatLng(36.9697908, 127.8750296),
+    LatLng(36.970273, 127.8755821),
+    LatLng(36.9710187, 127.8758047),
+    LatLng(36.9712394, 127.8763143),
+    LatLng(36.9718673, 127.8762848),
+    LatLng(36.9723623, 127.8760488),
+    LatLng(36.9730073, 127.8750725),
+    LatLng(36.9735752, 127.8749035),
+    LatLng(36.9745373, 127.8718216),
+    LatLng(36.9749209, 127.8718511),
+    LatLng(36.9749573, 127.871717),
+    LatLng(36.9749423, 127.8714327),
+    LatLng(36.9748866, 127.8712932),
+    LatLng(36.9748759, 127.8710518),
+    LatLng(36.9748266, 127.8709875),
+    LatLng(36.9747516, 127.8709794),
+    LatLng(36.974503, 127.871202),
+    LatLng(36.9744816, 127.8713871),
+    LatLng(36.9742309, 127.871481),
+    LatLng(36.9738688, 127.8714676),
+    LatLng(36.9734873, 127.8712637),
+    LatLng(36.9734595, 127.8712047),
+    LatLng(36.9732323, 127.8709392),
+    LatLng(36.9727823, 127.870899),
+    LatLng(36.9726773, 127.8708185),
+    LatLng(36.9709823, 127.8689517),
+    LatLng(36.9708559, 127.8688551),
+    LatLng(36.9696494, 127.8683455),
+    LatLng(36.9691351, 127.8682275),
+    LatLng(36.9687129, 127.8682677),
+    LatLng(36.9678493, 127.867742),
+    LatLng(36.9675857, 127.8676401),
+    LatLng(36.9674914, 127.8676428),
+    LatLng(36.9667563, 127.8711806),
+    LatLng(36.9664756, 127.87246),
+    LatLng(36.9664906, 127.8746728),
+)
+
+val KOREA_UNIVERSITY_POLYLINE = persistentListOf(
+    LatLng(37.5864637, 127.0306317),
+    LatLng(37.5865445, 127.0326487),
+    LatLng(37.5895158, 127.0356742),
+    LatLng(37.5904935, 127.0361141),
+    LatLng(37.592402, 127.0361785),
+    LatLng(37.59247, 127.0357171),
+    LatLng(37.5924573, 127.0355455),
+    LatLng(37.5915774, 127.034022),
+    LatLng(37.5911268, 127.0325253),
+    LatLng(37.5909568, 127.0316563),
+    LatLng(37.5912204, 127.0309482),
+    LatLng(37.5911523, 127.0305619),
+    LatLng(37.5912119, 127.0301435),
+    LatLng(37.5917644, 127.0293067),
+    LatLng(37.5921895, 127.0290599),
+    LatLng(37.592351, 127.0285449),
+    LatLng(37.5929036, 127.0282445),
+    LatLng(37.5929971, 127.027633),
+    LatLng(37.5930226, 127.0272896),
+    LatLng(37.5932266, 127.0270536),
+    LatLng(37.5937367, 127.0268283),
+    LatLng(37.5942553, 127.0266674),
+    LatLng(37.5945188, 127.0263562),
+    LatLng(37.5945528, 127.026088),
+    LatLng(37.5941958, 127.0253477),
+    LatLng(37.594013, 127.0252941),
+    LatLng(37.5935497, 127.0247791),
+    LatLng(37.5933202, 127.0245216),
+    LatLng(37.5928526, 127.0241461),
+    LatLng(37.5924785, 127.0239154),
+    LatLng(37.5921555, 127.0238189),
+    LatLng(37.5917772, 127.0238189),
+    LatLng(37.591367, 127.023784),
+    LatLng(37.591146, 127.0237679),
+    LatLng(37.5909951, 127.0237947),
+    LatLng(37.5907995, 127.0238484),
+    LatLng(37.5906401, 127.0239557),
+    LatLng(37.5905424, 127.0240978),
+    LatLng(37.5904637, 127.0243687),
+    LatLng(37.5904149, 127.0245136),
+    LatLng(37.5903107, 127.0246047),
+    LatLng(37.5902342, 127.0245913),
+    LatLng(37.5900737, 127.0245109),
+    LatLng(37.5899558, 127.0244961),
+    LatLng(37.589943, 127.0244941),
+    LatLng(37.5898899, 127.0245156),
+    LatLng(37.5897836, 127.0245806),
+    LatLng(37.5895137, 127.0248502),
+    LatLng(37.5892714, 127.0251157),
+    LatLng(37.5890525, 127.0251828),
+    LatLng(37.5889685, 127.0252257),
+    LatLng(37.5889281, 127.0252632),
+    LatLng(37.588925, 127.0252793),
+    LatLng(37.5889058, 127.0253638),
+    LatLng(37.5888729, 127.0254926),
+    LatLng(37.5888484, 127.0255784),
+    LatLng(37.5887974, 127.0256696),
+    LatLng(37.5887411, 127.0257393),
+    LatLng(37.5886497, 127.0257956),
+    LatLng(37.5883256, 127.0259351),
+    LatLng(37.5881194, 127.0260102),
+    LatLng(37.5878304, 127.025852),
+    LatLng(37.5874117, 127.0255784),
+    LatLng(37.5870716, 127.0254577),
+    LatLng(37.5869143, 127.0254175),
+    LatLng(37.586553, 127.0253504),
+    LatLng(37.5862193, 127.024712),
+    LatLng(37.5858962, 127.0244224),
+    LatLng(37.5854116, 127.0244063),
+    LatLng(37.585029, 127.0238966),
+    LatLng(37.5846336, 127.0238376),
+    LatLng(37.5841363, 127.024138),
+    LatLng(37.5837324, 127.0242185),
+    LatLng(37.5835071, 127.0243633),
+    LatLng(37.5835624, 127.0246638),
+    LatLng(37.5831585, 127.0248354),
+    LatLng(37.5830352, 127.0251948),
+    LatLng(37.5826101, 127.0258386),
+    LatLng(37.5824443, 127.0260853),
+    LatLng(37.5821978, 127.026257),
+    LatLng(37.5819384, 127.0261229),
+    LatLng(37.5817301, 127.0261497),
+    LatLng(37.5816536, 127.0263643),
+    LatLng(37.5816876, 127.0265145),
+    LatLng(37.5817939, 127.0266647),
+    LatLng(37.5821552, 127.0267237),
+    LatLng(37.5822743, 127.0269383),
+    LatLng(37.5823763, 127.0276517),
+    LatLng(37.5826186, 127.0278502),
+    LatLng(37.5825208, 127.0281721),
+    LatLng(37.5825804, 127.0282526),
+    LatLng(37.5827462, 127.0280433),
+    LatLng(37.5833498, 127.0281184),
+    LatLng(37.5833923, 127.0285208),
+    LatLng(37.5836091, 127.0288507),
+    LatLng(37.5853946, 127.02884),
+    LatLng(37.5854371, 127.0285315),
+    LatLng(37.5854626, 127.0284913),
+    LatLng(37.5858282, 127.0285261),
+    LatLng(37.5860003, 127.027979),
+    LatLng(37.5861491, 127.0279334),
+    LatLng(37.5864127, 127.0283464),
+    LatLng(37.5866465, 127.0286951),
+    LatLng(37.5873862, 127.0286576),
+    LatLng(37.5882278, 127.0287058),
+    LatLng(37.5886954, 127.0278314),
+    LatLng(37.5898134, 127.027354),
+    LatLng(37.5905402, 127.0274774),
+    LatLng(37.5908505, 127.0276705),
+    LatLng(37.5908973, 127.0280192),
+    LatLng(37.5907825, 127.0282767),
+    LatLng(37.590519, 127.0301596),
+    LatLng(37.59035, 127.030393),
+    LatLng(37.5904191, 127.0304332),
+    LatLng(37.5905275, 127.0305525),
+    LatLng(37.5908176, 127.0308637),
+    LatLng(37.590842, 127.0308704),
+    LatLng(37.5910912, 127.0307919),
+    LatLng(37.5911486, 127.0308241),
+    LatLng(37.5911794, 127.0308885),
+    LatLng(37.5911922, 127.0309502),
+    LatLng(37.5910392, 127.0313069),
+    LatLng(37.5909945, 127.0313565),
+    LatLng(37.5909212, 127.0313659),
+    LatLng(37.5908617, 127.0313391),
+    LatLng(37.5902677, 127.0309475),
+    LatLng(37.5900913, 127.0306095),
+    LatLng(37.5901423, 127.0304808),
+    LatLng(37.5898553, 127.0303091),
+    LatLng(37.5896556, 127.0304111),
+    LatLng(37.5895429, 127.030454),
+    LatLng(37.5892666, 127.0303172),
+    LatLng(37.5885344, 127.0303507),
+    LatLng(37.5883006, 127.0303802),
+    LatLng(37.5882028, 127.0305331),
+    LatLng(37.587952, 127.0305572),
+    LatLng(37.5878075, 127.0303936),
+    LatLng(37.5873824, 127.0304204),
+    LatLng(37.5868532, 127.0299162),
+    LatLng(37.586443, 127.030179),
+    LatLng(37.5864579, 127.0306216),
+)
+
+val SANGMYUNG_UNIVERSITY_POLYLINE = persistentListOf(
+    LatLng(126.9533132, 37.6013767),
+    LatLng(126.9534447, 37.6014298),
+    LatLng(126.9530584, 37.6016954),
+    LatLng(126.9526373, 37.6020864),
+    LatLng(126.9525864, 37.6024137),
+    LatLng(126.9528465, 37.6026687),
+    LatLng(126.9529753, 37.6028557),
+    LatLng(126.9532757, 37.6031),
+    LatLng(126.9533535, 37.603304),
+    LatLng(126.9534769, 37.6034783),
+    LatLng(126.9535627, 37.6036908),
+    LatLng(126.9536619, 37.6040903),
+    LatLng(126.9539033, 37.6041625),
+    LatLng(126.9540267, 37.6041753),
+    LatLng(126.9545176, 37.6041625),
+    LatLng(126.9548206, 37.6041285),
+    LatLng(126.9550084, 37.6041668),
+    LatLng(126.9551881, 37.6043793),
+    LatLng(126.9554027, 37.6042837),
+    LatLng(126.9556038, 37.6040435),
+    LatLng(126.9556119, 37.6039245),
+    LatLng(126.9557205, 37.6038937),
+    LatLng(126.9559485, 37.6041126),
+    LatLng(126.9566754, 37.6043899),
+    LatLng(126.9567693, 37.6045089),
+    LatLng(126.956831, 37.6045429),
+    LatLng(126.9569919, 37.6044664),
+    LatLng(126.9570455, 37.6044685),
+    LatLng(126.9570683, 37.6044738),
+    LatLng(126.9571609, 37.6044143),
+    LatLng(126.9575458, 37.6037928),
+    LatLng(126.9575029, 37.603491),
+    LatLng(126.9575672, 37.6033317),
+    LatLng(126.9572722, 37.6032594),
+    LatLng(126.9571998, 37.6031595),
+    LatLng(126.9570777, 37.6030735),
+    LatLng(126.9570013, 37.6029555),
+    LatLng(126.9569275, 37.6028408),
+    LatLng(126.9568819, 37.6027165),
+    LatLng(126.9568994, 37.6026538),
+    LatLng(126.956729, 37.6025762),
+    LatLng(126.956501, 37.6025699),
+    LatLng(126.9564125, 37.6026708),
+    LatLng(126.9562355, 37.6026017),
+    LatLng(126.9565212, 37.6022607),
+    LatLng(126.9565547, 37.601961),
+    LatLng(126.9565131, 37.6019408),
+    LatLng(126.9564756, 37.6019515),
+    LatLng(126.9563441, 37.6016965),
+    LatLng(126.9559512, 37.6017868),
+    LatLng(126.9554603, 37.6019865),
+    LatLng(126.9554067, 37.6019993),
+    LatLng(126.9553477, 37.6019918),
+    LatLng(126.9552297, 37.6019483),
+    LatLng(126.9549226, 37.6017942),
+    LatLng(126.9554483, 37.6014393),
+    LatLng(126.9545095, 37.6005192),
+    LatLng(126.9544854, 37.6004661),
+    LatLng(126.9544746, 37.6004299),
+    LatLng(126.954126, 37.6004639),
+    LatLng(126.9540079, 37.6006169),
+    LatLng(126.9535868, 37.6007211),
+    LatLng(126.9535037, 37.6009315),
+    LatLng(126.9533159, 37.6013735),
+)
